Moved CustomCookieJar to util.api.commmon package
